requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
markitdown==0.0.1a2
anthropic==0.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
pycryptodome==3.19.0